ui/featuers: Extract Plus API into its own slice
